image_to_relief/requirements.txt
numpy>=1.24
Pillow>=10.0
numpy-stl>=3.0
trimesh>=3.23
scipy>=1.10
